feat: changed styles, added proper drawer, loclization to mapview
