+++ src/components/TopHeader.tsx
@@ -1,16 +1,10 @@
-import { FontAwesome5 } from '@expo/vector-icons';
-import MapLibreGL from '@maplibre/maplibre-react-native';
 import { DrawerActions } from '@react-navigation/native';
-import { Stack, useNavigation } from 'expo-router';
-import { Drawer } from 'expo-router/drawer';
-import React, { useRef, useState } from 'react';
-import { View, Text, TextInput, TouchableOpacity } from 'react-native';
+import { useNavigation } from 'expo-router';
+import React from 'react';
+import { View, Text, TouchableOpacity } from 'react-native';
 
-import LightGreenDot from '../../assets/ellipse1.svg';
-import DarkGreenDot from '../../assets/ellipse2.svg';
 import MenuIcon from '../../assets/menus1.svg';
 import NavigationIcon from '../../assets/navigation.svg';
-import SearchIcon1 from '../../assets/search1.svg';
 import SearchIcon2 from '../../assets/search2.svg';
 
 type TopBarProps = {
@@ -18,18 +12,18 @@ type TopBarProps = {
   toggleSearchBar: () => void;
 };
 
-export default function TopBar({ isExpanded, toggleSearchBar }: TopBarProps) {
+export default function TopHeader({ isExpanded, toggleSearchBar }: TopBarProps) {
   const navigation = useNavigation();
   console.log('rerender', isExpanded);
   return (
-    <View className="absolute left-0 right-0 top-0 z-20 h-28 flex-row items-center justify-between bg-[#0F4530] px-4 pt-4">
+    <View className="absolute left-0 right-0 top-0 z-20 h-28 flex-row items-center justify-between bg-[#0F4530] px-4 pb-0 pt-8">
       <TouchableOpacity
         className="p-2"
         onPress={() => navigation.dispatch(DrawerActions.toggleDrawer())}>
         <MenuIcon width={40} height={40} fill="#FFF" />
       </TouchableOpacity>
       <Text className="text-lg font-bold text-white">logo/nazwa</Text>
-      <TouchableOpacity className="p-2" onPress={toggleSearchBar}>
+      <TouchableOpacity className="" onPress={toggleSearchBar}>
         {isExpanded ? (
           <NavigationIcon width={40} height={40} fill="#FFF" />
         ) : (
